requirements.txt
botorch==0.6.6
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
cycler==0.11.0
fonttools==4.38.0
gpytorch==1.6.0
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.5.3
multipledispatch==1.0.0
numpy==1.21.6
opt-einsum==3.3.0
packaging==24.0
Pillow==9.5.0
pyparsing==3.1.2
pyro-api==0.1.2
pyro-ppl==1.8.6
python-dateutil==2.9.0.post0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.1.0
torch==1.13.1
tqdm==4.66.4
typing_extensions==4.7.1
